requirements.txt
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
certifi==2024.8.30
charset-normalizer==3.4.1
click==8.1.7
comm==0.2.1
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.16.1
fonttools==4.48.1
fsspec==2025.2.0
h11==0.14.0
idna==3.8
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.2
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.5
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2025.1
PyYAML==6.0.2
pyzmq==25.1.2
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
sympy==1.13.3
threadpoolctl==3.2.0
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2025.1
ultralytics==8.3.72
ultralytics-thop==2.0.14
urllib3==2.2.2
uvicorn==0.33.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
zipp==3.17.0
